MEX-359: fixes after review
